class choice and parameter class tree, also some fixes on configure process

+++ app/Http/Controllers/Predicates/PredicateController.php
@@ -23,6 +23,14 @@ class PredicateController extends Controller
     public function delete($id)
     {
         $predicate = Predicate::find($id);
+
+        if ($predicate->instances->isNotEmpty()) {
+            return view('error')
+                ->with([
+                    'error_text' => 'Невозможно удалить ресурс, пока он связан с другими',
+                ]);
+        }
+
         $predicate->delete();
 
         return redirect()->route('web.predicates.index');
